+++ src/main/java/com/part4/team09/otboo/config/AsyncConfig.java
@@ -1,10 +1,23 @@
 package com.part4.team09.otboo.config;
 
+import java.util.concurrent.Executor;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
 @Configuration
 @EnableAsync
 public class AsyncConfig {
-  
+
+  @Bean(name = "recommendationExecutor")
+  public Executor recommendationExecutor() {
+    ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+    executor.setCorePoolSize(2);
+    executor.setMaxPoolSize(4);
+    executor.setQueueCapacity(20);
+    executor.setThreadNamePrefix("recommendation-");
+    executor.initialize();
+    return executor;
+  }
 }
